apps/proxy-server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
sse-starlette==1.8.2
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
redis[hiredis]==5.0.1
python-dotenv==1.0.0
email-validator==2.1.0
slowapi==0.1.9
pytest==7.4.4
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.11
aiofiles==23.2.1